Bug 1220309 - Correct all caps button text in Button from AppCompat. r=sebastian
AppCompat capitalizes all text in `Button`s so we have to override
that behavior to maintain the same UI. Ideally, we do this through
`android:buttonStyle` but the place I found the issue doesn't inherit
from that style so we can't and we change the style directly.

There may be issues with other `Button`s, but this is the only one I found.

UltraBlame original commit: f774157cf5f423be9a096ed5072b4440d68f4bd1