Bug 1219469 - Part 2: Make HttpChannelParentListener be the controller; r=jdm,jduell
The controller needs to be able to handle redirects.
HttpChannelParentListener is a listener object that can attach to
multiple HttpChannelParent objects as part of a single logical Necko
channel (meaning that the listener object will be reused for redirected
channels) so it seems like the better choice for the object that should
implement nsINetworkInterceptController.  This patch moves the
implementation of the nsINetworkInterceptController interface to
HttpChannelParentListener.

UltraBlame original commit: 7f9dd4660809cedac9938155ddf7a76ff85ceaf9