bug 1203861 - use strncmp() instead of nsCString.equals() to compare names r=davidb
AtkObject stores the name as a C string, but nsCString can store strings with
'\0' in the middle.  That means that if we compute a name containing '\0'
nsCString.equals() will nevr consider that string to be equal to
AtkObject::name, however we only want to compare up to the first '\0' because
that is the only part Atk will ever see.  So we should use strncmp() instead of
nsCString.equals().

UltraBlame original commit: 9133117f410c06ad7fa25323a168bc97200043b9