Bug 1221187: testing/marionette/client: Explicitly define --version flag; r=automatedtester
Passing the version keyword to argparse's constructor implicitly adds
-v and --version flags that conflicts with the -v/--verbose flag that
is added in BaseMarionetteArguments.

This removes -v as an alias to --version, which unbreaks testing on try
because it uses runtests.py.

It's perhaps worth mentioning that try uses runtests.py because it does
not (yet) have a full checkout of the source when it runs the tests.

UltraBlame original commit: 0dbc316b2160b6946dc2bcc73c4ada026ab1cc6d